Time: 47 ms (6.98%), Space: 41.1 MB (9.3%) - LeetHub

+++ 1889-check-if-number-is-a-sum-of-powers-of-three/1889-check-if-number-is-a-sum-of-powers-of-three.java
@@ -1,5 +1,5 @@
 class Solution {
-    public boolean helper(double n, int exp) {
+    public boolean helper(double n, double exp) {
         if (n == 0) return true;
         if (n < 0 || n < Math.pow(3, exp)) return false;
         return helper(n - Math.pow(3, exp), exp + 1) || helper(n, exp + 1);
